+++ packages/components/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.1.1...@ttoss/components@2.1.2) (2025-01-19)
+
+### Bug Fixes
+
+- notifications card theme ([#651](https://github.com/ttoss/ttoss/issues/651)) ([a93cf22](https://github.com/ttoss/ttoss/commit/a93cf22fb50f5da9248ffaef94589ead346d4cb5))
+
+## [2.1.1](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.1.0...@ttoss/components@2.1.1) (2025-01-18)
+
+### Bug Fixes
+
+- react-auth notifications ([#650](https://github.com/ttoss/ttoss/issues/650)) ([6eed312](https://github.com/ttoss/ttoss/commit/6eed3128d17633d7d9555f3dcd7705f050107c6c))
+
+# [2.1.0](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.0.24...@ttoss/components@2.1.0) (2025-01-18)
+
+### Features
+
+- react notifications v2 ([#649](https://github.com/ttoss/ttoss/issues/649)) ([5cc3237](https://github.com/ttoss/ttoss/commit/5cc3237ecdb48b03b46c1b0fe4bef43d2e3f5046))
+
 ## [2.0.24](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.0.23...@ttoss/components@2.0.24) (2025-01-13)
 
 **Note:** Version bump only for package @ttoss/components