+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.28.9](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.28.8...@ttoss/forms@0.28.9) (2025-01-19)
+
+**Note:** Version bump only for package @ttoss/forms
+
+## [0.28.8](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.28.7...@ttoss/forms@0.28.8) (2025-01-18)
+
+**Note:** Version bump only for package @ttoss/forms
+
 ## [0.28.7](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.28.6...@ttoss/forms@0.28.7) (2025-01-13)
 
 **Note:** Version bump only for package @ttoss/forms