+++ terezinha-farm/vite-app/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.13](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.10.12...@terezinha-farm/vite-app@0.10.13) (2025-01-19)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
+## [0.10.12](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.10.11...@terezinha-farm/vite-app@0.10.12) (2025-01-18)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
+## [0.10.11](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.10.10...@terezinha-farm/vite-app@0.10.11) (2025-01-18)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
 ## [0.10.10](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.10.9...@terezinha-farm/vite-app@0.10.10) (2025-01-13)
 
 **Note:** Version bump only for package @terezinha-farm/vite-app